+++ src/fmmax/basis.py
@@ -157,12 +157,14 @@ def _reciprocal(lattice_vectors: LatticeVectors) -> LatticeVectors:
     cross_product = _cross_product(lattice_vectors.u, lattice_vectors.v)
     uprime = (
         jnp.stack([lattice_vectors.v[..., 1], -lattice_vectors.v[..., 0]], axis=-1)
-        / cross_product
+        / cross_product[..., jnp.newaxis]
     )
     vprime = (
         jnp.stack([-lattice_vectors.u[..., 1], lattice_vectors.u[..., 0]], axis=-1)
-        / cross_product
+        / cross_product[..., jnp.newaxis]
     )
+    assert uprime.shape == lattice_vectors.u.shape
+    assert vprime.shape == lattice_vectors.v.shape
     return LatticeVectors(u=uprime, v=vprime)
 
 
@@ -304,13 +306,18 @@ def transverse_wavevectors(
     """
     reciprocal_vectors = primitive_lattice_vectors.reciprocal
     kx = in_plane_wavevector[..., 0, jnp.newaxis] + 2 * jnp.pi * (
-        expansion.basis_coefficients[:, 0] * reciprocal_vectors.u[..., 0]
-        + expansion.basis_coefficients[:, 1] * reciprocal_vectors.v[..., 0]
+        expansion.basis_coefficients[:, 0] * reciprocal_vectors.u[..., 0, jnp.newaxis]
+        + expansion.basis_coefficients[:, 1] * reciprocal_vectors.v[..., 0, jnp.newaxis]
     )
     ky = in_plane_wavevector[..., 1, jnp.newaxis] + 2 * jnp.pi * (
-        expansion.basis_coefficients[:, 0] * reciprocal_vectors.u[..., 1]
-        + expansion.basis_coefficients[:, 1] * reciprocal_vectors.v[..., 1]
+        expansion.basis_coefficients[:, 0] * reciprocal_vectors.u[..., 1, jnp.newaxis]
+        + expansion.basis_coefficients[:, 1] * reciprocal_vectors.v[..., 1, jnp.newaxis]
     )
+    batch_shape = jnp.broadcast_shapes(
+        in_plane_wavevector.shape[:-1],
+        primitive_lattice_vectors.u.shape[:-1],
+    )
+    assert kx.shape == batch_shape + (expansion.num_terms,)
     return jnp.stack([kx, ky], axis=-1)