+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## Unreleased
+- Add permittivity shape validation to eigensolve functions.
+- Make small changes throughout to acommodate primitive lattice vectors with batch dimensions.s
 
 ## 1.2.0 (April 2, 2025)
 - In fields module, allow number of unit cells to be specified independently when Brillouin zone integration is used. Retain ability to infer number of unit cells from Brillouin grid axes in when number of unit cells is not specified.